Changes to measures script

+++ scripts/measures.py
@@ -1,6 +1,6 @@
 #! /usr/bin/env python
 
-import numpy as np
+import os.path
 import pandas as pd
 
 match_info = ["Season",
@@ -8,6 +8,7 @@ match_info = ["Season",
               "Date",
               "HomeTeam",
               "AwayTeam"]
+
 match_statistics = ["FTHG",
                     "FTAG",
                     "FTR",
@@ -35,18 +36,22 @@ match_statistics = ["FTHG",
                     "HBP",
                     "ABP"]
 
-def create_measures_from_historical_scores(df):
-    measures = df[match_info].copy()
-    measures["Date"] = pd.to_datetime(measures["Date"])
-    measures.to_csv("../data/processed/measures.csv")
-    return measures
+def get_measures():
+    if os.path.isfile("../data/processed/measures.csv"):
+        measures = pd.read_csv("../data/processed/measures.csv")
+        return measures
+    else:
+        measures = pd.read_csv("../data/processed/historical_scores.csv", index_col=0)
+        measures = measures[match_info]
+        measures["Date"] = pd.to_datetime(measures["Date"])
+        measures.to_csv("../data/processed/measures.csv")
+        return measures
 
 def main():
-    historical_scores = pd.read_csv("../data/processed/historical_scores.csv", index_col=0)
-    historical_scores = historical_scores[match_info + match_statistics]
+    measures = get_measures()
 
-    measures = create_measures_from_historical_scores(historical_scores)
-
+    raw_data = pd.read_csv("../data/processed/historical_scores.csv", index_col=0)
+    raw_data = raw_data[match_info + match_statistics]
 
 if __name__ == "__main__":
     main()
